finished configuring basic rollup build and watch functionality
