+++ README.md
@@ -1,7 +1,9 @@
 # ear-rhythm-training
 Web app that tests your musical ear by playing short random sporadic sequences of sound, and validating it by tapping the spacebar at the exact rhythm.
 
+
 Run with:
 
-cd mysite
-python manage.py runserver
+"python mysite/manage.py runserver"
+
+To edit JS, run "npm start" to make live changes to JS app
