--- mysite/music_training/static/js/components/input.js
@@ -1,3 +1,0 @@
-
-
-//export input as
