+++ src/components/instruments.js
@@ -1,4 +1,4 @@
-export function startKeyboard(tempo) {
+function keyboard(tempo) {
     let synth = new Tone.Synth().toDestination();
 
     const reverb = new Tone.Reverb({
@@ -35,3 +35,19 @@ export function startKeyboard(tempo) {
 
     return synth;
 }
+
+function fourByFour() {
+    const kickDrum = new Tone.MembraneSynth({
+        volume: 4
+    }).toDestination();
+
+    const kickPart = new Tone.Part(function(time) {
+        kickDrum.triggerAttackRelease('C1', '8n', time);
+    }, [{ time: '0:0' },{ time: '0:1' },{ time: '0:2' },{ time: '0:3' }]).start(0);
+
+    kickPart.loop = true;
+
+    return kickPart;
+}
+
+export { keyboard, fourByFour };
