--- src/components/test.js
@@ -1,11 +1,0 @@
-function test() {
-    alert('hierfg');
-}
-
-function launch(included) {
-    included();
-}
-
-launch(test);
-alert("I live at the end of the file!");
-alert("I live of the file!xf");
